Refactor Ruler, add MultiTenantManager interface
